[FEAT] Community 관련 기능 대폭 추가 CommunityService class createPost(게시글 작성) getCertainPost(특정 게시글 fetch) getPostsWithLimitAndIdAfter(게시판에서 게시글 fetch) createThreadOfTrhead(대댓글 작성) createThread(댓글 작성) 등등의 기능을 추가하였고, 그에 맞는 테스트 작성 필요

+++ community/src/main/java/webdoc/community/domain/entity/post/Picture.java
@@ -13,7 +13,7 @@ import webdoc.community.domain.BaseEntity;
 @EqualsAndHashCode(of = "id")
 public class Picture extends BaseEntity {
 
-    public static Picture createPicture(String address, Long priority){
+    public static Picture createPicture(String address, Integer priority){
         Picture picture = new Picture();
         picture.setAddress(address);
         picture.setPriority(priority);
@@ -31,7 +31,7 @@ public class Picture extends BaseEntity {
     private Post post;
 
     @Column(nullable = false)
-    private Long priority;
+    private Integer priority;
 
 
 }
